+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 
 Added
 -----
+- Allow staff users to create new ``Variant`` objects through the API endpoint
 
 Changed
 -------